requirements.txt
flask
psycopg2-binary
flask-sqlalchemy
Flask-Migrate
flask_validator
flask-script
flask_socketio==5.3.3
python-socketio>=5.0.2
python-dateutil
datetime